AWSCHAT-76 make form component reususable
