+++ src/components/TollChat/TollChat.module.scss
@@ -371,90 +371,6 @@
     }
 }
 
-.form {
-    width: 100%;
-    z-index: 3;
-    background: #fff;
-    right: 10px;
-    bottom: 0;
-    padding: 35px 15px;
-    color: #000;
-    font-family: "Gotham", sans-serif;
-    
-    input {
-        border: 2px solid #d0d7db;
-        padding: 20px;
-        width: 100%;
-        color: #000;
-        font-size: 1em;
-        font-family: $alt_font_family;
-        font-weight: $font_weight_medium;
-        margin-bottom: 15px;
-        margin: 0 0 15px;
-
-        &::placeholder {
-            color: rgba(0,0,0,.8);
-            font-family: $special_font_family;
-            font-size: 14px;
-        }
-
-    }
-
-    button[type=submit] {
-        background-color: #1463c2;
-        padding: 18px;
-        color: #fff;
-        width: 100%;
-        font-size: 1em;
-        font-family: $special_font_family;
-        border: none;
-        text-align: center;
-    }
-
-    .privacyPolicy {
-        color: black;
-        margin: 20px 0 30px 0;
-        font-size: .8em;
-        line-height: 1.3;
-        text-align: left;
-
-        a {
-            text-decoration: underline;
-        }
-    }
-
-    .agent {
-
-        label, .radioLabel {
-            font-size: .875em;
-            font-family: $special_font_family;
-            color: #000;
-        }
-
-        .radioGroup {
-            display: flex;
-            gap: 12px;
-
-            label {
-                display: inline-flex;
-                align-items: center;
-                gap: 4px;
-                font-size: .875em;
-                color: #000;
-
-                input {
-                    width: auto;
-                    margin: 0;
-
-                    &:focus-visible {
-                        box-shadow: inset 0 0 1px 1px #0070cd;
-                    }
-                }
-            }
-        }
-    }
-}
-
 .chatInputWrapper {
     position: relative;
     padding: 0 15px;
